+++ startiosdev/Assets.xcassets/SideMenu.imageset/Contents.json
@@ -6,12 +6,12 @@
     },
     {
       "idiom" : "universal",
-      "filename" : "一键水印.png",
+      "filename" : "SideMenu.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "一键水印@3x.png",
+      "filename" : "SideMenu@3x.png",
       "scale" : "3x"
     }
   ],
