+++ startiosdev/PWMNavigationVC.swift
@@ -14,7 +14,7 @@ class PWMNavigationVC: UINavigationController {
         super.viewDidLoad()
 
         self.navigationBar.barTintColor = UIColor(red: 32/255, green: 32/255, blue: 32/255, alpha: 0.2)
-        
+        self.navigationBar.tintColor = UIColor.whiteColor()
         //UIApplication.sharedApplication().setStatusBarStyle(.Default, animated: true)
         //self.navigationBar.translucent = true
     }
